% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract all the relevant data for analysis from a fit-object.}
\usage{
extract_fit(fit)
}
\description{
Extract all the relevant data for analysis from a fit-object.
}
