% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fit.R
\name{import_fit}
\alias{import_fit}
\title{Funktion, die FIT-Dateien importiert.}
\usage{
import_fit(device = "garmin", from = NULL, to = NULL)
}
\arguments{
\item{device}{character. Auswahl zwischen Garmin und Lezyne Gerät. Damit wird
das Quellen- und das Speicherverzeichnis bestimmt.}

\item{from}{character. Pfad zum Quellenverzeichnis. Muss angegeben werden,
wenn das default-Verzeichnis der Geräte nicht mit dem tatsächlichen
Quellenverzeichnis übereinstimmt. Zum Beispiel wenn das Gerät aktuell nicht
als das default-Laufwerk ('E:/') angezeigt wird oder wenn Aktivitäten von
einem anderen Gerät importiert werden sollen. Kann nur in Kombination mit dem
Argument 'to' übergeben werden.}

\item{to}{character. Pfad zum Speicherverzeichnis. Siehe 'from'.}
}
\description{
Funktion, die FIT-Dateien importiert.
}
