% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_to_pace.R
\name{speed_to_pace}
\alias{speed_to_pace}
\title{Rechnet m/s in s/km um.}
\usage{
speed_to_pace(speed)
}
\arguments{
\item{speed}{numeric. Geschwindigkeit in [m/s].}
}
\description{
Rechnet m/s in s/km um.
}
